bugfix: método onClick funcionando perfeitamente (95% do sistema de progress está concluído)
