+++ app/src/main/java/com/android/parasitologymobilesoftware/HomeActivity.java
@@ -249,141 +249,6 @@ public class HomeActivity extends AppCompatActivity
         datePickerDialog.show();
     }
 
-    public void onCategoryButtonClick(View view) {
-        if (true) { //@Todo Arrumar isso ae
-            Intent intent = new Intent(HomeActivity.this, CategoryActivity.class);
-            categoryId = view.getId();
-
-            switch (view.getId()) {
-                case R.id.categoryHomeFragmentIntroduction:
-                    category = "Introdução";
-                    break;
-                case R.id.categoryHomeFragmentIntroductionEcologia:
-                    category = "Ecologia";
-                    break;
-                case R.id.categoryHomeFragmentIntroductionConceitosGerais:
-                    category = "Conceitos Gerais";
-                    break;
-                case R.id.categoryHomeFragmentIntroductionClassificacao:
-                    category = "Classificação";
-                    break;
-                case R.id.categoryHomeFragmentIntroductionReproducao:
-                    category = "Reprodução";
-                    break;
-                case R.id.categoryHomeFragmentIntroductionCicloBiologico:
-                    category = "Ciclo Biológico";
-                    break;
-                case R.id.categoryHomeFragmentIntroductionAtualidades:
-                    category = "Atualidades";
-                    break;
-                case R.id.categoryHomeFragmentProtozoarios:
-                    category = "Protozoários";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosAmebiase:
-                    category = "Amebíase";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosGiardiase:
-                    category = "Giardíase";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosLeishmanioses:
-                    category = "Leishmanioses";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosTricomonose:
-                    category = "Tricomonose";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosChagas:
-                    category = "Doença de Chagas";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosMalaria:
-                    category = "Malária";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosToxoplasmose:
-                    category = "Toxoplasmose";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosBalantidiase:
-                    category = "Balantidíase";
-                    break;
-                case R.id.categoryHomeFragmentProtozoariosProtozooses:
-                    category = "Protozooses";
-                    break;
-                case R.id.categoryHomeFragmentHelmintos:
-                    category = "Helmintos";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosEsquistossomose:
-                    category = "Esquistossomose";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosFascioliase:
-                    category = "Fasciolíase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosTeniase:
-                    category = "Teníase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosCisticercose:
-                    category = "Cisticercose";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosHidatidose:
-                    category = "Hidatidose";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosHimenolepiase:
-                    category = "Himenolepíase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosEstrongiloidiase:
-                    category = "Estrongiloidíase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosTricuriase:
-                    category = "Tricuríase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosAncilostomiase:
-                    category = "Ancilostomíase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosNecatoriase:
-                    category = "Necatoríase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosEnterobiase:
-                    category = "Enterobíase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosAscaridiase:
-                    category = "Ascaridíase";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosLarvaMigrans:
-                    category = "Larva Migrans";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosFilarioses:
-                    category = "Filarioses";
-                    break;
-                case R.id.categoryHomeFragmentHelmintosOutrasHelmintoses:
-                    category = "Outras Helmintoses";
-                    break;
-                case R.id.categoryHomeFragmentArtropodes:
-                    category = "Artrópodes";
-                    break;
-                case R.id.categoryHomeFragmentArtropodesHemipteros:
-                    category = "Hemípteros";
-                    break;
-                case R.id.categoryHomeFragmentArtropodesMosquitos:
-                    category = "Mosquitos";
-                    break;
-                case R.id.categoryHomeFragmentArtropodesMoscas:
-                    category = "Moscas";
-                    break;
-                case R.id.categoryHomeFragmentArtropodesEctoparasitos:
-                    category = "Ectoparasitos";
-                    break;
-            }
-            Bundle extra = new Bundle();
-            extra.putString("category", category);
-            extra.putInt("CategoryId", categoryId);
-
-            // Passing category's name to activity
-            intent.putExtras(extra);
-            // Passing category's name to fragment
-            categoryFragment.setArguments(extra);
-
-            startActivity(intent);
-        }
-
-    }
-
     public void settingProgress(int progressToGo){                      // Call this function to change the progress bar, passing an int variable that goes from 0 to 100
         DocumentReference docRef = dataBase.collection("generalUserInfo").document(email);
         docRef.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
